Added lots of changes
Yo Jason! Have made a double jump power up, in krita, a better software than paint. Imported that but didn't yet put it in scene. I increased the speed but I think it's too fast so will let you change it accordingly. The best part is that when you die, you see ui popup that says "Menu" and anotherone that says "restart". Both work well. Also the main menu now looks good and works. Your task is to implement the double jump powerup and change the speed. Commit to main when you do it :) Best Of Luck!!